refactor of code model impl
