+++ src/generate/generators/CoreIncre.ts
@@ -7,7 +7,7 @@ import * as path from 'path';
 import { SystemType, PathConstants, AzConfiguration, CodeGenConstants } from '../../utils/models';
 import { thoughtAsTrue } from '../../utils/helper';
 import { GeneratorBase } from './Base';
-import { CodeModelAz } from '../CodeModelAz';
+import { CodeModelAz } from '../codemodel/CodeModelAz';
 import { GenerateNamespaceInit } from '../renders/CliNamespaceInit';
 import { CliReport } from '../renders/CliReport';
 import { CliTopAction } from '../renders/CliTopAction';
