+++ src/generate/renders/CliMeta.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *-------------------------------------------------------------------------------------------- */
 
-import { CodeModelAz } from '../CodeModelAz';
+import { CodeModelAz } from '../codemodel/CodeModelAz';
 import { putToZip } from '../../utils/inplace';
 
 export function GenerateMetaFile(model: CodeModelAz) {
