+++ src/generate/renders/TemplateBase.ts
@@ -2,7 +2,7 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *-------------------------------------------------------------------------------------------- */
-import { CodeModelAz } from '../CodeModelAz';
+import { CodeModelAz } from '../codemodel/CodeModelAz';
 import * as nunjucks from 'nunjucks';
 import { AzConfiguration, CodeGenConstants } from '../../utils/models';
