+++ src/generate/renders/extraExt/CliExtSetupCfg.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *-------------------------------------------------------------------------------------------- */
 
-import { CodeModelAz } from '../../CodeModelAz';
+import { CodeModelAz } from '../../codemodel/CodeModelAz';
 import * as path from 'path';
 import { TemplateBase } from '../TemplateBase';
 import { PathConstants } from '../../../utils/models';
